Move tas-client to umd version and adopt in web (#106904)
* adopt tas-client-umd

* adopt in web

* fix whitespace

* adding tas-client to webignore

* upgrade tas-client-umd

* move svc imort to common

* remove unnecessary ignore
