+++ cglicenses.json
@@ -348,7 +348,7 @@
 	},
 	{
 		// Reason: The license cannot be found by the tool due to access controls on the repository
-		"name": "tas-client",
+		"name": "tas-client-umd",
 		"fullLicenseText": [
 			"MIT License",
 			"Copyright (c) 2020 - present Microsoft Corporation",
