+++ package.json
@@ -63,7 +63,7 @@
     "semver-umd": "^5.5.7",
     "spdlog": "^0.11.1",
     "sudo-prompt": "9.1.1",
-    "tas-client": "^0.0.950",
+    "tas-client-umd": "^0.1.1",
     "v8-inspect-profiler": "^0.0.20",
     "vscode-nsfw": "1.2.8",
     "vscode-oniguruma": "1.3.1",
