+++ remote/web/package.json
@@ -5,6 +5,7 @@
 		"iconv-lite-umd": "0.6.8",
 		"jschardet": "2.2.1",
 		"semver-umd": "^5.5.7",
+		"tas-client-umd": "0.1.1",
 		"vscode-oniguruma": "1.3.1",
 		"vscode-textmate": "5.2.0",
 		"xterm": "4.10.0-beta.4",
